+++ example/lib/test.dart
@@ -1,6 +1,3 @@
-import 'package:simple_json_mapper/simple_json_mapper.dart';
-
-@JsonObject()
 class Test {
   const Test({this.name});
   final String name;