+++ simple_json/lib/simple_json.dart
@@ -2,4 +2,4 @@ library simple_json;
 
 import 'package:simple_json/src/json_mapper_builder.dart';
 
-JsonMapperBuilder jsonMapperBuilder(_) => const JsonMapperBuilder();
+JsonMapperBuilder jsonMapperBuilder(_) => JsonMapperBuilder();