+++ simple_json/lib/src/json_mapper_builder.dart
@@ -6,10 +6,14 @@ import 'package:simple_json_mapper/simple_json_mapper.dart';
 import 'package:source_gen/source_gen.dart';
 import 'package:path/path.dart' as p;
 
+import 'utils.dart';
+
 class JsonMapperBuilder implements Builder {
-  const JsonMapperBuilder();
+  JsonMapperBuilder();
 
   static final _allFilesInLib = new Glob('lib/**');
+
+  final processedTypes = Set<DartType>();
 
   @override
   Map<String, List<String>> get buildExtensions {
@@ -32,15 +36,32 @@ class JsonMapperBuilder implements Builder {
     await for (final input in buildStep.findAssets(_allFilesInLib)) {
       final library = await buildStep.resolver.libraryFor(input);
       classesInLibrary.addAll(LibraryReader(library)
-          .annotatedWithExact(TypeChecker.fromRuntime(JsonObject))
+          .annotatedWith(TypeChecker.fromRuntime(JsonObject))
           .where((match) => match.element is ClassElement)
           .map((match) => match.element as ClassElement)
           .toList());
     }
 
-    lines.add(_generateHeader(classesInLibrary));
-    lines.addAll(classesInLibrary.map((c) => _generateMapper(c)));
-    lines.add(_generateInit(classesInLibrary));
+    final mappers = classesInLibrary.map((c) => _generateMapper(c)).toList();
+    final unmappedTypes = processedTypes
+        .where((t) => t.element is ClassElement)
+        .toSet()
+        .difference(classesInLibrary
+            .map((optedClasses) => optedClasses.thisType)
+            .toSet());
+    print(
+        'WARNING: Generated mappings for the following unmarked (but linked) types: ${unmappedTypes.toString()}');
+    final unmappedElements =
+        unmappedTypes.map((t) => t.element as ClassElement).toList();
+    mappers.addAll(unmappedElements.map((c) => _generateMapper(c)));
+
+    final classes = [...classesInLibrary, ...unmappedElements];
+    final imports = _generateHeader(classes);
+    final registrations = _generateInit(classes);
+
+    lines.add(imports);
+    lines.addAll(mappers);
+    lines.add(registrations);
 
     await buildStep.writeAsString(_allFileOutput(buildStep), lines.join('\n'));
   }
@@ -52,19 +73,28 @@ class JsonMapperBuilder implements Builder {
 
 final _${elementName.toLowerCase()}Mapper = JsonObjectMapper(
   (Map<String, dynamic> json) => ${elementName}(
-    ${parameters.map(_generateFromMapParameter).join('\n    ')}
+    ${parameters.map(_generateFromMapItem).where((line) => line != null).join('\n    ')}
   ),
   (${elementName} instance) => <String, dynamic>{
-    ${parameters.map(_generateToMapItem).join('\n    ')}
+    ${parameters.map(_generateToMapItem).where((line) => line != null).join('\n    ')}
   },
 );
 ''';
   }
 
-  String _generateFromMapParameter(ParameterElement param) {
+  String _generateFromMapItem(ParameterElement param) {
+    final prop = getProp(param);
+    if (prop.ignore) return null;
     final name = param.displayName;
+    final jsonName = prop.name ?? name;
     final type = param.type.toString();
-    final valFn = ([String asType]) => '''json['$name'] as ${asType ?? type}''';
+    final valFn = ([String asType]) {
+      var accessorStr = "json['$jsonName']";
+      if (prop.defaultValue != null) {
+        accessorStr = '($accessorStr ?? ${prop.defaultValue.toString()})';
+      }
+      return '''$accessorStr as ${asType ?? type}''';
+    };
     var val;
     switch (type) {
       case 'DateTime':
@@ -76,19 +106,38 @@ final _${elementName.toLowerCase()}Mapper = JsonObjectMapper(
     if (param.type.isDartCoreList) {
       final typeArgs = (param.type as InterfaceType).typeArguments;
       if (typeArgs.isNotEmpty && !isPrimitiveType(typeArgs.first)) {
+        processedTypes.add(typeArgs.first);
         val =
             '''(${valFn('List')}).cast<Map<String, dynamic>>().map((item) => ${_generateDeserialize('item', typeArgs.first)}).toList()''';
       } else {
         val = '''(${valFn('List')}).cast()''';
       }
+    } else if (isParamFieldFormal(param) && isParamEnum(param)) {
+      final enumProp = getEnumProp(param);
+      final enumValueMap = cleanMap(getEnumValueMap(param, enumProp));
+      val = _generateEnumFromMap(param, enumProp, enumValueMap);
     } else if (!isPrimitiveType(param.type) && val == null) {
+      processedTypes.add(param.type);
       val = _generateDeserialize(valFn('Map<String, dynamic>'), param.type);
     }
     return '''$name: ${val ?? valFn()},''';
   }
 
+  String _generateEnumFromMap(ParameterElement param, JsonEnumProperty enumProp,
+      Map<dynamic, dynamic> enumValueMap) {
+    final value = enumProp.serializationType == SerializationType.Index
+        ? 'item.index'
+        : "item.toString().split('.')[1]";
+    return '''${param.type}.values.firstWhere(
+        (item) => ${_generateMapLookup(enumValueMap, value)} == json['${param.displayName}'],
+        orElse: () => null)''';
+  }
+
   String _generateToMapItem(ParameterElement param) {
+    final prop = getProp(param);
+    if (prop.ignore) return null;
     final name = param.displayName;
+    final jsonName = prop.name ?? name;
     final valFn = () => 'instance.${name}';
     var val;
     switch (param.type.toString()) {
@@ -101,13 +150,101 @@ final _${elementName.toLowerCase()}Mapper = JsonObjectMapper(
     if (param.type.isDartCoreList) {
       final typeArgs = (param.type as InterfaceType).typeArguments;
       if (typeArgs.isNotEmpty && !isPrimitiveType(typeArgs.first)) {
+        processedTypes.add(typeArgs.first);
         val =
             '''${valFn()}.map((item) => ${_generateSerialize('item', typeArgs.first)}).toList()''';
       }
+    } else if (isParamFieldFormal(param) && isParamEnum(param)) {
+      final enumProp = getEnumProp(param);
+      final enumValueMap = cleanMap(getEnumValueMap(param, enumProp));
+      val = _generateMapLookup(
+          enumValueMap, '${valFn()}${_generateEnumToMap(param, enumProp)}');
     } else if (!isPrimitiveType(param.type) && val == null) {
+      processedTypes.add(param.type);
       val = _generateSerialize(valFn(), param.type);
     }
-    return ''''$name': ${val ?? valFn()},''';
+    return ''''$jsonName': ${val ?? valFn()}${prop.defaultValue != null ? ' ?? ${prop.defaultValue.toString()}' : ''},''';
+  }
+
+  String _generateEnumToMap(ParameterElement param, JsonEnumProperty enumProp) {
+    return '''${enumProp.serializationType == SerializationType.Index ? '.index' : ".toString().split('.')[1]"}''';
+  }
+
+  String _generateMapLookup(Map<dynamic, dynamic> map, String val) {
+    return map.isNotEmpty ? '''(${map.toString()}[${val}] ?? ${val})''' : val;
+  }
+
+  JsonProperty getProp(ParameterElement param) {
+    final propChecker = TypeChecker.fromRuntime(JsonProperty);
+    final jsonPropType = param.isInitializingFormal
+        ? propChecker
+            .firstAnnotationOf((param as FieldFormalParameterElement).field)
+        : null;
+    final fieldMap = jsonPropType?.getField('(super)') ?? jsonPropType;
+    final jsonProp = JsonProperty(
+      ignore: fieldMap?.getField('ignore')?.toBoolValue() ?? false,
+      name: fieldMap?.getField('name')?.toStringValue(),
+      defaultValue:
+          ConstantReader(fieldMap?.getField('defaultValue'))?.literalValue,
+    );
+    return jsonProp;
+  }
+
+  JsonEnumProperty getEnumProp(ParameterElement param) {
+    final propChecker = TypeChecker.fromRuntime(JsonEnumProperty);
+    final jsonPropType = param.isInitializingFormal
+        ? propChecker
+            .firstAnnotationOf((param as FieldFormalParameterElement).field)
+        : null;
+    final fieldMap = jsonPropType?.getField('(super)') ?? jsonPropType;
+    final jsonProp = JsonEnumProperty(
+      serializationType: fieldMap != null
+          ? SerializationType.values.firstWhere((val) =>
+              fieldMap
+                  ?.getField('serializationType')
+                  ?.getField(Utils.enumToString(val)) !=
+              null)
+          : SerializationType.Value,
+    );
+    return jsonProp;
+  }
+
+  Map<dynamic, dynamic> getEnumValueMap(
+      ParameterElement param, JsonEnumProperty enumProp) {
+    final propChecker = TypeChecker.fromRuntime(EnumValue);
+    final enumElement = ((param as FieldFormalParameterElement)
+        .field
+        .type
+        .element as ClassElement);
+    final enumFields =
+        enumElement.fields.where((e) => e.isEnumConstant).toList();
+    final enumValProps = enumFields
+        .map((field) => propChecker.firstAnnotationOf(field))
+        .toList();
+    return enumFields.asMap().entries.fold(<dynamic, dynamic>{},
+        (map, fieldEntry) {
+      final displayName = fieldEntry.value.displayName;
+      final key = enumProp.serializationType == SerializationType.Index
+          ? ConstantReader(enumElement
+                  .getField(displayName)
+                  .computeConstantValue()
+                  .getField(displayName))
+              .intValue
+          : displayName;
+      final value =
+          ConstantReader(enumValProps[fieldEntry.key].getField('value'))
+                  .literalValue ??
+              key;
+      map[key] = value;
+      return map;
+    });
+  }
+
+  Map<dynamic, dynamic> cleanMap(Map<dynamic, dynamic> valueMap) {
+    return valueMap.entries.fold(<dynamic, dynamic>{}, (map, entry) {
+      if (entry.key != entry.value) map[entry.key] = entry.value;
+      return map;
+    });
   }
 
   String _generateSerialize(String val, DartType type) {
@@ -124,6 +261,14 @@ final _${elementName.toLowerCase()}Mapper = JsonObjectMapper(
         type.isDartCoreInt ||
         type.isDartCoreNum ||
         type.isDartCoreString;
+  }
+
+  bool isParamFieldFormal(ParameterElement param) {
+    return param.isInitializingFormal;
+  }
+
+  bool isParamEnum(FieldFormalParameterElement param) {
+    return (param.field.type.element as ClassElement).isEnum;
   }
 
   String _generateInit(List<ClassElement> elements) {
@@ -144,11 +289,17 @@ void init() {
       '''// Generated and consumed by 'simple_json' ''',
       '',
       '''import 'package:simple_json_mapper/simple_json_mapper.dart';''',
-      elements.map(_generateImport).join('\n')
+      Utils.dedupe(elements.map(_generateImport).toList()).join('\n')
     ].join('\n');
   }
 
   String _generateImport(ClassElement element) {
     return '''import '${element.library.identifier}';''';
   }
+
+  // T toObjectOfType<T>(DartObject dartObject, ParameterizedType type) {
+  //   type.con
+  //   return
+  // }
+
 }