+++ simple_json_mapper/lib/simple_json_mapper.dart
@@ -1,4 +1,4 @@
 library simple_json_mapper;
 
-export 'src/annotations.dart' show JsonObject;
-export 'src/json_mapper.dart' show JsonMapper, JsonObjectMapper;
+export 'src/annotations.dart';
+export 'src/json_mapper.dart';