+++ simple_json_mapper/lib/src/annotations.dart
@@ -1,3 +1,50 @@
+class JObj extends JsonObject {
+  const JObj() : super();
+}
+
 class JsonObject {
   const JsonObject();
+}
+
+class JProp extends JsonProperty {
+  const JProp({
+    bool ignore,
+    String name,
+    dynamic defaultValue,
+  }) : super(
+          ignore: ignore,
+          name: name,
+          defaultValue: defaultValue,
+        );
+}
+
+class JsonProperty {
+  const JsonProperty({
+    this.ignore,
+    this.name,
+    this.defaultValue,
+  });
+  final bool ignore;
+  final String name;
+  final dynamic defaultValue;
+}
+
+enum SerializationType {
+  Value,
+  Index,
+}
+
+class JsonEnumProperty {
+  const JsonEnumProperty({this.serializationType});
+  final SerializationType serializationType;
+}
+
+class JEnumProp extends JsonEnumProperty {
+  const JEnumProp({SerializationType serializationType})
+      : super(serializationType: serializationType);
+}
+
+class EnumValue {
+  const EnumValue({this.value});
+  final dynamic value;
 }