+++ simple_json_mapper/CHANGELOG.md
@@ -1,3 +1,18 @@
+## 0.1.6 - 06/15/2020
+* Fix unintentionally hidden annotations.
+
+## 0.1.5 - 06/15/2020
+* Add enum annotations. `JsonEnumProperty` (with alias `JEnumProp`) and `EnumValue`.
+
+## 0.1.4 - 06/15/2020
+* Fix bug `JObj` const constructor.
+
+## 0.1.3 - 06/15/2020
+* Fix bug with not revealing the new annotations.
+
+## 0.1.2 - 06/15/2020
+* Add `JObj` as an alias to `JsonObject` and `JsonProperty` (with `JProp` alias) for controlling class property options.
+
 ## 0.1.1 - 06/15/2020
 * Add example project.